requirements.txt
fastapi
uvicorn
requests
google-generativeai
slowapi
python-dotenv
